requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
sqlalchemy
python-jose[cryptography]
passlib[argon2]
python-multipart
python-dotenv
pydantic
pydantic[email]
